use <a> instead of <button> for back button
...and change style of the back button.

Also, the back button now works in firefox.

In firefox, it seems that an <a> element doesn't trigger default
behavior of click under <button> element. Also, the order by which
the default behavior and the defined handler for <a> is called is
undefined; and if defined handler changes href of <a> first, the
default behavior of changing hash is prevented, i.e., the page
fails to jump to the destination.

Now in addition to click event, hashchange event is handled. So
there is no undefined behavior.