CSS print media: float left 条标(.title)
It appears that "page-break-after: avoid" on .title elements isn't
respected in either Firefox or Chrome. While the rule is respected
in Edge, I still look for a way to mitigate the problem of having
a .title element appear on the bottom of a page while its following
paragraph is on the next page.

The idea came from print: let the first paragraph of a 条 appear
on the same line as the 条标(title).

This style allows browsers to take the first paragraph of each 条
as an integral part of the title of the 条. Thus, we no longer need
browsers' support of "page-break-after: avoid" on .title elements
to mitigate the discontinuity of contents across pages.

Two side effects of this style change:
 - The .title elements are now less obvious given that they no
   longer appear at a standalone line serving as an extra separater
   between 条s in addition to the vertical margin given to each 条.
   (We still apply bold font to them, and by not apply text indent
   on this line, they are still special and should be clear to
   readers.)
 - A line is saved for each 条. Thus printing the entire documents
   uses less pages of paper. Good cost saving both economically and
   environmentally.